requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-static==0.0.6
Django==4.1.2
django-4-jet==1.0.9
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.5
whitenoise==6.2.0
